moves ViewUpdaterSpy into a separate file
